Reformatted to fit inside a single file
Moved the Pattern class inside the text class (now called Match).
First: Match is instantiated with the text to be searched.  Next: patterns are added to the Match object. To get pattern matches, the Match object is run through the match_patterns function specifying the desired patterns by name (all as default).